Add diagnostic FEN benchmark test and reusable probe
